Split sound module
