+++ core/src/mmu.rs
@@ -1,3 +1,4 @@
+use crate::apu::Apu;
 use crate::cgb::Cgb;
 use crate::cpu::Sys;
 use crate::dma::{Dma, DmaRequest};
@@ -8,7 +9,6 @@ use crate::ic::{Ic, Irq};
 use crate::joypad::Joypad;
 use crate::mbc::Mbc;
 use crate::serial::Serial;
-use crate::sound::Sound;
 use crate::timer::Timer;
 use crate::wram::Wram;
 use alloc::vec::Vec;
@@ -28,7 +28,7 @@ pub struct Mmu {
     ic: Ic,
     serial: Serial,
     joypad: Joypad,
-    sound: Sound,
+    apu: Apu,
     dma: Dma,
     cgb: Cgb,
 }
@@ -47,7 +47,7 @@ impl Mmu {
             ic: Ic::new(irq.clone()),
             serial: Serial::new(hw.clone(), irq.clone()),
             joypad: Joypad::new(hw.clone(), irq),
-            sound: Sound::new(hw),
+            apu: Apu::new(hw),
             dma: Dma::new(),
             cgb: Cgb::new(color),
         }
@@ -62,28 +62,28 @@ impl Mmu {
             0xff04..=0xff07 => self.timer.on_read(addr),
             0xff08..=0xff0e => todo!("i/o read: addr={:04x}", addr),
             0xff0f => self.ic.read_flags(),
-            0xff10 => self.sound.read_tone_sweep(),
-            0xff11 => self.sound.read_tone_wave(0),
-            0xff12 => self.sound.read_tone_envelop(0),
-            0xff13 => self.sound.read_tone_freq_low(0),
-            0xff14 => self.sound.read_tone_freq_high(0),
-            0xff16 => self.sound.read_tone_wave(1),
-            0xff17 => self.sound.read_tone_envelop(1),
-            0xff18 => self.sound.read_tone_freq_low(1),
-            0xff19 => self.sound.read_tone_freq_high(1),
-            0xff1a => self.sound.read_wave_enable(),
-            0xff1b => self.sound.read_wave_len(),
-            0xff1c => self.sound.read_wave_amp(),
-            0xff1d => self.sound.read_wave_freq_low(),
-            0xff1e => self.sound.read_wave_freq_high(),
-            0xff20 => self.sound.read_noise_len(),
-            0xff21 => self.sound.read_noise_envelop(),
-            0xff22 => self.sound.read_noise_poly_counter(),
-            0xff23 => self.sound.read_noise_select(),
-            0xff24 => self.sound.read_ctrl(),
-            0xff25 => self.sound.read_so_mask(),
-            0xff26 => self.sound.read_enable(),
-            0xff30..=0xff3f => self.sound.read_wave_buf(addr),
+            0xff10 => self.apu.read_tone_sweep(),
+            0xff11 => self.apu.read_tone_wave(0),
+            0xff12 => self.apu.read_tone_envelop(0),
+            0xff13 => self.apu.read_tone_freq_low(0),
+            0xff14 => self.apu.read_tone_freq_high(0),
+            0xff16 => self.apu.read_tone_wave(1),
+            0xff17 => self.apu.read_tone_envelop(1),
+            0xff18 => self.apu.read_tone_freq_low(1),
+            0xff19 => self.apu.read_tone_freq_high(1),
+            0xff1a => self.apu.read_wave_enable(),
+            0xff1b => self.apu.read_wave_len(),
+            0xff1c => self.apu.read_wave_amp(),
+            0xff1d => self.apu.read_wave_freq_low(),
+            0xff1e => self.apu.read_wave_freq_high(),
+            0xff20 => self.apu.read_noise_len(),
+            0xff21 => self.apu.read_noise_envelop(),
+            0xff22 => self.apu.read_noise_poly_counter(),
+            0xff23 => self.apu.read_noise_select(),
+            0xff24 => self.apu.read_ctrl(),
+            0xff25 => self.apu.read_so_mask(),
+            0xff26 => self.apu.read_enable(),
+            0xff30..=0xff3f => self.apu.read_wave_buf(addr),
             0xff40 => self.gpu.read_ctrl(),
             0xff41 => self.gpu.read_status(),
             0xff42 => self.gpu.read_scy(),
@@ -126,28 +126,28 @@ impl Mmu {
             0xff04..=0xff07 => self.timer.on_write(addr, v),
             0xff08..=0xff0e => todo!("i/o write: addr={:04x}, v={:02x}", addr, v),
             0xff0f => self.ic.write_flags(v),
-            0xff10 => self.sound.write_tone_sweep(v),
-            0xff11 => self.sound.write_tone_wave(0, v),
-            0xff12 => self.sound.write_tone_envelop(0, v),
-            0xff13 => self.sound.write_tone_freq_low(0, v),
-            0xff14 => self.sound.write_tone_freq_high(0, v),
-            0xff16 => self.sound.write_tone_wave(1, v),
-            0xff17 => self.sound.write_tone_envelop(1, v),
-            0xff18 => self.sound.write_tone_freq_low(1, v),
-            0xff19 => self.sound.write_tone_freq_high(1, v),
-            0xff1a => self.sound.write_wave_enable(v),
-            0xff1b => self.sound.write_wave_len(v),
-            0xff1c => self.sound.write_wave_amp(v),
-            0xff1d => self.sound.write_wave_freq_low(v),
-            0xff1e => self.sound.write_wave_freq_high(v),
-            0xff20 => self.sound.write_noise_len(v),
-            0xff21 => self.sound.write_noise_envelop(v),
-            0xff22 => self.sound.write_noise_poly_counter(v),
-            0xff23 => self.sound.write_noise_select(v),
-            0xff24 => self.sound.write_ctrl(v),
-            0xff25 => self.sound.write_so_mask(v),
-            0xff26 => self.sound.write_enable(v),
-            0xff30..=0xff3f => self.sound.write_wave_buf(addr, v),
+            0xff10 => self.apu.write_tone_sweep(v),
+            0xff11 => self.apu.write_tone_wave(0, v),
+            0xff12 => self.apu.write_tone_envelop(0, v),
+            0xff13 => self.apu.write_tone_freq_low(0, v),
+            0xff14 => self.apu.write_tone_freq_high(0, v),
+            0xff16 => self.apu.write_tone_wave(1, v),
+            0xff17 => self.apu.write_tone_envelop(1, v),
+            0xff18 => self.apu.write_tone_freq_low(1, v),
+            0xff19 => self.apu.write_tone_freq_high(1, v),
+            0xff1a => self.apu.write_wave_enable(v),
+            0xff1b => self.apu.write_wave_len(v),
+            0xff1c => self.apu.write_wave_amp(v),
+            0xff1d => self.apu.write_wave_freq_low(v),
+            0xff1e => self.apu.write_wave_freq_high(v),
+            0xff20 => self.apu.write_noise_len(v),
+            0xff21 => self.apu.write_noise_envelop(v),
+            0xff22 => self.apu.write_noise_poly_counter(v),
+            0xff23 => self.apu.write_noise_select(v),
+            0xff24 => self.apu.write_ctrl(v),
+            0xff25 => self.apu.write_so_mask(v),
+            0xff26 => self.apu.write_enable(v),
+            0xff30..=0xff3f => self.apu.write_wave_buf(addr, v),
             0xff40 => self.gpu.write_ctrl(v),
             0xff41 => self.gpu.write_status(v),
             0xff42 => self.gpu.write_scy(v),
